+++ includes/class-rest-api.php
@@ -211,7 +211,7 @@ class ODW_Rest_API {
 
 		if ( ! empty( $theme ) ) {
 			$meta_query[] = array(
-				'key'   => '_odw_theme',
+				'key'   => '_odw_theme_index',
 				'value' => class_exists( 'ODW_Fields' ) ? ODW_Fields::resolve_theme_uri( $theme ) : $theme,
 			);
 		}